+++ src/uploads/getImageSize.ts
@@ -1,13 +1,10 @@
-import { UploadedFile } from 'express-fileupload';
+import fs from 'fs';
 import probeImageSize from 'probe-image-size';
+import { UploadedFile } from 'express-fileupload';
+import { ProbedImageSize } from './types';
 
-export type ProbedImageSize = {
-  width: number,
-  height: number,
-  type: string,
-  mime: string,
-}
-
-export default async function (image: UploadedFile): Promise<ProbedImageSize> {
-  return probeImageSize.sync(image.data);
+export default async function (file: UploadedFile): Promise<ProbedImageSize> {
+  return file.tempFilePath
+    ? probeImageSize(fs.createReadStream(file.tempFilePath))
+    : probeImageSize.sync(file.data);
 }